avoid exception

+++ candidates/cec_api.py
@@ -46,9 +46,11 @@ for candidate in [(u'桃園市', 5, u'羅文欽'), (u'桃園市', 3, u'黃志浩
         delete from candidates_terms
         where ad = 9 and county = %s and constituency = %s and name = %s returning candidate_id
     ''', (candidate[0], candidate[1], candidate[2]))
-    candidate_id = c.fetchone()[0]
-    c.execute('''
-        delete from candidates_candidates
-        where uid = %s
-    ''', [candidate_id])
+    res = c.fetchone()
+    if res:
+        candidate_id = res[0]
+        c.execute('''
+            delete from candidates_candidates
+            where uid = %s
+        ''', [candidate_id])
 conn.commit()
